Comprehensive CI implementation (#1)
* Comprehensive CI implementation

* Resolved incorrect clang-14 apt dependency; cppcheck; doc output path

* lcov, genhtml, pip, and some other dependencies for CI

* clang-format applied; algorithm include fix

* Required C++20 in Installation Test job

* sign-compare warnings fixed

* modified CI for clang / gcc14 compatibility issues

* .githooks; cppcheck fix

* coverage flag off by default

* githook format fix; cppcheck: use compile_commands.json

* autoformat githook

* githooks fix; cppcheck fix

* githook fix for diff-tree --root

* Fixed integration test

* Specify clang-format-14

* clang-format issues

* clang-format not working right...try this diagnostic + fix...

* clang-format --style=file fix

* clang-format again...

* Explicitly force CI to find .clang-format file...

* Debugging CI-side of clang-format

* Full clang-format dump

* Force clang-format-14

* FINALLY figured out the clang-format issues

* Resolve Warnings as Errors clang CI job libc++ dependency

+++ examples/example_master.cpp
@@ -89,7 +89,7 @@ int main() {
 
   // Example 8: Broadcast write (slave ID 0)
   std::cout << "\nExample 8: Broadcast write to all slaves (slave ID 0)...\n";
-  success = master.WriteSingleRegister(0, 0, 0xABCD);  // Broadcast
+  success = master.WriteSingleRegister(0, 0, static_cast<int16_t>(0xABCD));  // Broadcast
   if (success) {
     std::cout << "  Broadcast write successful (no response expected)\n";
   }
@@ -98,8 +98,7 @@ int main() {
   std::cout << "\nExample 9: Reading exception status from slave 1...\n";
   auto exception_status = master.ReadExceptionStatus(1);
   if (exception_status.has_value()) {
-    std::cout << "  Exception status: 0x" << std::hex
-              << static_cast<int>(*exception_status) << std::dec << "\n";
+    std::cout << "  Exception status: 0x" << std::hex << static_cast<int>(*exception_status) << std::dec << "\n";
   }
 
   // Example 10: Read FIFO queue
